Refactor ResourceBundle.Control implementation to be more efficient and extract it to its own class, with a backing custom ResourceBundle implementation.
