+++ hapi-fhir-base/src/main/java/ca/uhn/fhir/i18n/HapiLocalizer.java
@@ -23,28 +23,15 @@ import ca.uhn.fhir.context.ConfigurationException;
 import ca.uhn.fhir.util.UrlUtil;
 import ca.uhn.fhir.util.VersionUtil;
 import com.google.common.annotations.VisibleForTesting;
-import org.apache.commons.lang3.StringUtils;
 
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStreamWriter;
-import java.io.StringReader;
-import java.io.Writer;
-import java.net.URL;
-import java.nio.charset.StandardCharsets;
 import java.text.MessageFormat;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.Enumeration;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
-import java.util.Properties;
-import java.util.PropertyResourceBundle;
 import java.util.ResourceBundle;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
@@ -183,95 +170,7 @@ public class HapiLocalizer {
 	protected void init(String[] theBundleNames) {
 		myBundle = new ArrayList<>();
 		for (String nextName : theBundleNames) {
-			// LUKETODO:  this fails because we eventually get: Unknown localization key: moduletype.ADMIN_JSON.shortName
-//			final ResourceBundle bundleOldStyle = ResourceBundle.getBundle(nextName);
-//			myBundle.add(bundleOldStyle);
-//			ourLog.info("1234: Loading bundle: {}", nextName);
-			final ResourceBundle bundleNewStyle = ResourceBundle.getBundle(nextName, new MultiFileResourceBundleControl());
-//			if (! bundleOldStyle.keySet().equals(bundleNewStyle.keySet())) {
-//				ourLog.warn("1234: bundle keys do not match for class: {} {}: old style keys: class: {} {}", bundleOldStyle.getClass(), bundleOldStyle.getKeys(), bundleNewStyle.getClass(), bundleNewStyle.getKeys());
-//			}
-//			ourLog.info("1234: bundle keySet: {} contains dqm: {} or cql: {}", bundleNewStyle.keySet(), bundleNewStyle.containsKey("module_config.evaluate_measure.chunksize"), bundleNewStyle.containsKey("module_config.cql.use_embedded_libraries"));
-			myBundle.add(bundleNewStyle);
-		}
-	}
-
-	/**
-	 * Find all properties files on the class path that match the given base name and locale, and merge them into a single
-	 * properties object.
-	 */
-//	public static class MultiFileResourceBundleControl extends ResourceBundle.Control {
-	public static class MultiFileResourceBundleControl extends ResourceBundle.Control {
-		/**
-		 * Returns a list containing only Locale.ROOT as a candidate.
-		 * This forces the loader to look for the base bundle only (e.g., "bundle.properties").
-		 * @param baseName the base name of the resource bundle
-		 * @param locale   the locale to load (this parameter is ignored)
-		 * @return a singleton list containing Locale.ROOT
-		 */
-		@Override
-		public List<Locale> getCandidateLocales(String baseName, Locale locale) {
-			return Collections.singletonList(Locale.ROOT);
-		}
-
-		/**
-		 * Prevents fallback to the default locale if the root bundle isn't found.
-		 * @param baseName the base name of the resource bundle
-		 * @param locale   the locale where the search is failing
-		 * @return null to indicate no fallback should be attempted
-		 */
-		@Override
-		public Locale getFallbackLocale(String baseName, Locale locale) {
-			return null;
-		}
-
-		@Override
-		public ResourceBundle newBundle(
-				String baseName, Locale locale, String format, ClassLoader loader, boolean reload)
-				throws IOException, IllegalAccessException, InstantiationException {
-//			ourLog.info("1234: locale: {}", locale);
-			if ("java.properties".equals(format) && (locale == null || locale.toString().isEmpty())) {
-				String bundleName = toBundleName(baseName, locale);
-				String resourceName = toResourceName(bundleName, "properties");
-
-				Properties properties = new Properties();
-
-				var urls = new ArrayList<>();
-				// Load from all matching resources
-				Enumeration<URL> resources = loader.getResources(resourceName);
-//				final ArrayList<URL> list = Collections.list(resources);
-//				ourLog.info("1234: before loop");
-				while (resources.hasMoreElements()) {
-					URL url = resources.nextElement();
-					urls.add(url);
-					try (InputStream is = url.openStream()) {
-						properties.load(is);
-					}
-//					ourLog.info("1234: urL: {}, properties contains dqm: {} or cql: {}", url, properties.get("module_config.evaluate_measure.chunksize"), properties.get("module_config.cql.use_embedded_libraries"));
-				}
-//				ourLog.info("1234: after loop");
-
-//				ourLog.info("1234:resourceName:{}, urls: {} properties.toString(): {}", resourceName, urls, properties.toString());
-
-//				ourLog.info("1234: bundleName: {}, resourceName: {}, properties size: {} properties contains dqm: {} or cql: {}", bundleName, resourceName,
-//				properties.entrySet().size(), properties.get("module_config.evaluate_measure.chunksize"), properties.get("module_config.cql.use_embedded_libraries"));
-
-
-				ByteArrayOutputStream outStream = new ByteArrayOutputStream();
-				try (Writer writer = new OutputStreamWriter(outStream, StandardCharsets.UTF_8)) {
-					properties.store(writer, "module properties");
-				}
-
-				InputStream inStream = new ByteArrayInputStream(outStream.toByteArray());
-				final PropertyResourceBundle bundle = new PropertyResourceBundle(inStream);
-
-
-
-//				final PropertyResourceBundle bundle = new PropertyResourceBundle(properties.store()));
-//				ourLog.info("1234: bundle keySet: {} contains dqm: {} or cql: {}", bundle.keySet(), bundle.containsKey("module_config.evaluate_measure.chunksize"), bundle.containsKey("module_config.cql.use_embedded_libraries"));
-				return bundle;
-			}
-			return super.newBundle(baseName, locale, format, loader, reload);
+			myBundle.add(ResourceBundle.getBundle(nextName, new MultiFileResourceBundleControl()));
 		}
 	}
